+++ elastic/datadog_checks/elastic/__init__.py
@@ -2,6 +2,6 @@
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 from .__about__ import __version__
-from .elastic import ESCheck
+from .elastic import ESCheck, AuthenticationError
 
-__all__ = ['__version__', 'ESCheck']
+__all__ = ['__version__', 'ESCheck', 'AuthenticationError']